chore: refactor tools services
